fully factor out train_step
